+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+### Unreleased
+
+* Ensure `ZemeckisTestUtil.resetConfig(...)` cancels delayed and periodic tasks and waits for in-flight tasks to finish.
+
 ### [v0.15](https://github.com/realityforge/zemeckis/tree/v0.15) (2026-07-20) · [Full Changelog](https://github.com/realityforge/zemeckis/compare/v0.14...v0.15)
 
 Changes in this release: